requirements.txt
django>=4.2
Pillow>=10.0
